refactor: 암호화 기능 패키지 위치 변경
