+++ src/main/java/com/ghm/giftcardfleamarket/user/service/UserService.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
-import com.ghm.giftcardfleamarket.common.utils.encrypt.PasswordEncryptor;
+import com.ghm.giftcardfleamarket.common.utils.PasswordEncryptor;
 import com.ghm.giftcardfleamarket.user.domain.User;
 import com.ghm.giftcardfleamarket.user.dto.request.LoginRequest;
 import com.ghm.giftcardfleamarket.user.dto.request.SignUpRequest;
