+++ src/test/java/com/ghm/giftcardfleamarket/user/service/SessionLoginServiceTest.java
@@ -14,8 +14,8 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mock.web.MockHttpSession;
 
+import com.ghm.giftcardfleamarket.common.utils.PasswordEncryptor;
 import com.ghm.giftcardfleamarket.common.utils.constants.SessionNames;
-import com.ghm.giftcardfleamarket.common.utils.encrypt.PasswordEncryptor;
 import com.ghm.giftcardfleamarket.user.domain.User;
 
 @ExtendWith(MockitoExtension.class)
